Creating a websocket creation endpoint for ios
